refactor: add typescript
well, you cannot run from typescript
