+++ package.json
@@ -19,7 +19,7 @@
   "devDependencies": {
     "@eslint/js": "^9.17.0",
     "@react-router/dev": "^7.2.0",
-    "@types/react": "^18.3.17",
+    "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
     "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.20",
@@ -32,6 +32,7 @@
     "globals": "^15.13.0",
     "postcss": "^8.4.49",
     "tailwindcss": "^3.4.17",
+    "typescript": "^5.7.3",
     "vite": "^6.0.3"
   }
 }
